rename "helper" classes -> "service" classes
OCD-3093
