+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/domain/service/ChangeRequestDetailsFactory.java
@@ -1,4 +1,4 @@
-package gov.healthit.chpl.changerequest.manager;
+package gov.healthit.chpl.changerequest.domain.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -6,18 +6,18 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class ChangeRequestDetailsFactory {
-    private ChangeRequestWebsiteHelper crWebsiteHelper;
+    private ChangeRequestWebsiteService crWebsiteHelper;
 
     @Value("${changerequest.website}")
     private Long websiteChangeRequestType;
 
     @Autowired
-    public ChangeRequestDetailsFactory(final ChangeRequestWebsiteHelper crWebsiteHelper) {
+    public ChangeRequestDetailsFactory(final ChangeRequestWebsiteService crWebsiteHelper) {
         this.crWebsiteHelper = crWebsiteHelper;
     }
 
-    public ChangeRequestDetailsHelper<?> get(Long changeRequestType) {
-        ChangeRequestDetailsHelper<?> crDetailsHelper = null;
+    public ChangeRequestDetailsService<?> get(Long changeRequestType) {
+        ChangeRequestDetailsService<?> crDetailsHelper = null;
 
         if (changeRequestType.equals(changeRequestType)) {
             crDetailsHelper = crWebsiteHelper;
